// src/app/sections/Testimonials/Testimonials.module.css.d.ts
declare const styles: {
  'testimonials-section': string;
  wrapper: string;
  carousel: string;
  container: string;
  card: string;
  cardActive: string;
  cardPrev: string;
  cardNext: string;
  content: string;
  text: string;
  clientInfo: string;
  clientImage: string;
  clientDetails: string;
  clientName: string;
  clientPosition: string;
  nav: string;
  navButton: string;
};

export default styles; 